requirements.txt
flask
langchain
langchain-text-splitters
chromadb
openai
pypdf
python-docx
sentence-transformers